+++ userInput.c
@@ -1,4 +1,4 @@
-#include "userInput.h"
+#include "./includes/userInput.h"
 
 /*
 *	 helper method to flush the stream to the next line.
